+++ packages/attestations-common/src/extractors/index.ts
@@ -30,7 +30,7 @@ import {
   IBaseAttAssets,
   IBaseAttGender,
 } from '../AttestationData'
-import {TAttestationTypeNames} from '../AttestationTypes'
+import {TAttestationTypeNames} from '../types'
 
 export const extract = (dataStr: string, attType: TAttestationTypeNames, valType: string, errCallback?: (err: any) => void) => {
   const a: string | IBaseAtt | null = parseDataStr(dataStr)