+++ packages/attestations-common/src/index.ts
@@ -3,7 +3,6 @@ import * as EthUtils from './ethUtils'
 import * as Utils from './utils'
 
 export * from './extractors'
-export * from './AttestationTypes'
 export * from './RFC3339DateTime'
 export * from './types'
 export * from './validation'