+++ packages/attestations-common/src/types/index.ts
@@ -1,0 +1,5 @@
+export * from './auth'
+export * from './vc'
+export * from './vp'
+export * from './attestation'
+export * from './misc'