+++ packages/attestations-common/src/types/vc/full/index.ts
@@ -1,0 +1,6 @@
+import {FullVCV1} from './v1'
+
+export type FullVC = FullVCV1
+
+export * from './v0'
+export * from './v1'