+++ packages/attestations-common/src/types/vp/index.ts
@@ -1,0 +1,6 @@
+import {VerifiablePresentationV1} from './v1'
+
+export type VerifiablePresentation = VerifiablePresentationV1
+
+export * from './v0'
+export * from './v1'