+++ packages/attestations-common/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0
+
+**Improvements**
+
+- Improve file structure for types
+- Add new types that are more standardized verifiable credentials and verifiable presentations
+  - The way the types are set up allow us to make new versions without having to make breaking changes to this library
+
 ## 1.0.0
 
 - Initial release