+++ packages/verify-kit/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.1.0
+
+**Improvements**
+
+- Implement validation of new VC/VP types from [attestations-common][attestations-common]
+
 ## 2.0.0
 
 **Improvements**
@@ -8,8 +14,10 @@
 **Breaking**
 
 - Most functions have been removed from the public API. The two main entry points are now `validateVerifiablePresentationResponse` and `validateVerifiableAuthResponse`
-- All `type`s have been removed from this library, they are exported by [attestations-common](https://github.com/hellobloom/attestations-ts/tree/master/packages/attestations-common) now.
+- All `type`s have been removed from this library, they are exported by [attestations-common][attestation-common] now.
 
 ## 1.0.0
 
 - Initial release
+
+[attestation-common]: https://github.com/hellobloom/attestations-ts/tree/master/packages/attestations-common